src/test/java/com/assignment/employeeManagement/controller/EmployeeControllerTest.java
package com.assignment.employeeManagement.controller;

import com.assignment.employeeManagement.dto.EmployeeDTO;
import com.assignment.employeeManagement.entity.Employee;
import com.assignment.employeeManagement.service.EmployeeService;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import java.security.Principal;
import java.util.Arrays;
import java.util.List;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class EmployeeControllerTest {

    @Mock
    private EmployeeService employeeService;

    @InjectMocks
    private EmployeeController employeeController;

    @Test
    public void testFetchAllEmployee() {
        List<Employee> employees = Arrays.asList(new Employee(), new Employee());
        when(employeeService.getAllEmployees()).thenReturn(employees);

        ResponseEntity<List<Employee>> response = employeeController.fetchAllEmployee();

        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertEquals(employees, response.getBody());
    }

    @Test
    public void testUpdateSkills() {
        Principal principal = () -> "username";
        List<String> skills = Arrays.asList("Java", "Spring");
        Employee employee = new Employee();
        when(employeeService.updateSkills(any(Principal.class), any(List.class))).thenReturn(employee);

        ResponseEntity<Employee> response = employeeController.updateSkills(principal, skills);

        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertEquals(employee, response.getBody());
    }

    @Test
    public void testGetInfo() {
        Principal principal = () -> "username";
        Employee employee = new Employee();
        when(employeeService.getEmployeeInfo(any(Principal.class))).thenReturn(employee);

        ResponseEntity<Employee> response = employeeController.getInfo(principal);

        assertEquals(HttpStatus.OK, response.getStatusCode());
        assertEquals(employee, response.getBody());
    }

    
}
